// src/types/user.ts
export interface User {
  id: number;
  first_name: string;
  last_name: string;
  email: string;
  avatar: string;
}

export interface UserListState {
  users: User[];
  isLoading: boolean;
  error: string | null;
}

export interface UserListApiResponse {
  page: number;
  per_page: number;
  total: number;
  total_pages: number;
  data: User[];
}